refactor into separate class
